# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CoursesList API
  version: "1.0.0"
  summary: API systemu obsługi szkoleń — kursy, kursanci, firmy, zaświadczenia, dzienniki szkoleń.
  description: |
    Specyfikacja dla aplikacji integrujących się z API serwer-serwer za pomocą klucza API.

    **Przeczytaj najpierw `INTEGRATION.md`** z tego samego katalogu — opisuje reguły, których
    nie da się wyrazić w samym schemacie (semantyka PATCH jako pełnego nadpisania, formaty dat
    bez strefy czasowej, niespójności nazw pól, pułapki przy tworzeniu zaświadczeń).

    Konwencje:
    - Każda udana odpowiedź JSON ma kopertę `{"data": ...}`; błędy mają `{"error": {"code", "message"}}`.
    - Każda operacja wymaga zakresu podanego w `x-required-scope` (brak = wystarczy ważny klucz).
    - Operacje z `x-requires-admin: true` wymagają dodatkowo, by konto serwisowe klucza miało rolę
      administratora (`role = 1`) — sam zakres nie wystarczy.
    - Ciała żądań są odrzucane (400), jeśli zawierają pole spoza schematu.

    Wygenerowano z kodu źródłowego (`api/internal/*`) na gałęzi `api_and_webhook`.
servers:
  - url: "{baseUrl}/api/v1"
    variables:
      baseUrl:
        default: http://localhost:8081
        description: >-
          Adres Go API. Lokalnie `http://localhost:8081`; w `docker-compose.prod.yml` API nasłuchuje na
          porcie 8082. Integracje powinny wołać Go API bezpośrednio, nie przez proxy Nuxt.
security:
  - apiKey: []
...
paths:
...
  /admin/users:
    get:
      tags: [admin]
      operationId: listUsers
      summary: Lista użytkowników systemu
      x-required-scope: users:read
      x-requires-admin: true
      responses:
        "200":
          description: Użytkownicy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [admin]
      operationId: createUser
      summary: Utworzenie użytkownika
      x-required-scope: users:write
      x-requires-admin: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserCreate"
      responses:
        "201":
          description: Utworzony użytkownik.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "400":
          description: "`invalid request body` lub `invalid email format`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: "Adres e-mail jest zajęty (`user with this email already exists`)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: >-
        Klucz API w nagłówku `Authorization: Bearer clk_...`. Klucze wystawia administrator w aplikacji
        webowej (Administracja → Klucze API). Klucz działa w imieniu przypisanego konta serwisowego i
        tylko w zakresie nadanych uprawnień (`x-required-scope`).
...
  responses:
...
    Unauthorized:
      description: >-
        Brak lub nieważny klucz: `missing credentials`, `invalid or expired api key`.
        Klucz mógł zostać unieważniony, wygasnąć albo jego konto usunięto.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: >-
        Klucz jest ważny, ale nie ma uprawnień: `api key is missing the <zakres> scope`
        albo `admin access required` (konto serwisowe nie jest administratorem).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Błąd serwera (`internal_error`). Treść komunikatu nie jest stabilna.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, invalid_credentials, forbidden, not_found, conflict, internal_error]
            message:
              type: string
              description: >-
                Komunikat po angielsku. Komunikaty wymienione w opisach operacji są stabilne i można na nich
                opierać logikę; pozostałe traktuj jako informacyjne.
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, role]
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: "#/components/schemas/Role"
...
    Role:
      type: integer
      description: >-
        `1` = administrator. **Każda inna wartość** (w danych występują `0` i `2`) to zwykły użytkownik —
        sprawdzaj `role === 1`, nie zakładaj konkretnej wartości dla zwykłego konta.
      examples: [1, 2, 0]
    RoleInput:
      type: integer
      minimum: 1
      description: "`1` = administrator. Przy zapisie wymagana wartość dodatnia; aplikacja webowa dla zwykłego użytkownika używa `2`."
      examples: [1, 2]

    # -------------------------------------------------------------- students
...
    AdminUserListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuthUser"
    AdminUserResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AuthUser"
    AdminUserCreate:
      type: object
      additionalProperties: false
      required: [email, password, firstName, lastName, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
          description: Białe znaki na brzegach są obcinane. Brak wymogu minimalnej długości przy tworzeniu.
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        role:
          $ref: "#/components/schemas/RoleInput"
...